Wrap ClassMethods into NxtSupport::Services::Base; remove blocks support
